+++ src/OnchainKitProvider.tsx
@@ -1,10 +1,13 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { createContext, useMemo } from 'react';
 import { WagmiProvider } from 'wagmi';
-import { ONCHAIN_KIT_CONFIG, setOnchainKitConfig } from './OnchainKitConfig';
+import {
+  ONCHAIN_KIT_CONFIG,
+  setOnchainKitConfig,
+} from './core/utils/OnchainKitConfig';
 import { DEFAULT_PRIVACY_URL, DEFAULT_TERMS_URL } from './core/constants';
 import { checkHashLength } from './core/utils/checkHashLength';
-import { createWagmiConfig } from './createWagmiConfig';
+import { createWagmiConfig } from './core/utils/createWagmiConfig';
 import { COINBASE_VERIFIED_ACCOUNT_SCHEMA_ID } from './identity/constants';
 import type { OnchainKitContextType, OnchainKitProviderReact } from './types';
 import { useProviderDependencies } from './useProviderDependencies';
